Split scripts into modules
