+++ models/decoders/__init__.py
@@ -1,0 +1,2 @@
+from .decoder_transformer import GraphTransformerDecoder
+from .two_pointer_fcn import TwoPointerFCN
