+++ models/encoders/gnn/encoder_gatv2conv.py
@@ -8,7 +8,7 @@ from torch import Tensor
 import torch
 
 
-class GATv2ConvEncoder(pl.LightningModule):
+class TransConvEncoder(pl.LightningModule):
     def __init__(self, in_channels, hidden_channels, num_layers, edge_dim_) -> None:
         super().__init__()
         modules: list[Any] = []
