+++ models/encoders/gnn/encoder_gru.py
@@ -2,7 +2,7 @@ import pytorch_lightning as pl
 import torch
 
 
-class EncoderGRU(pl.LightningModule):
+class GRUEncoder(pl.LightningModule):
     def __init__(self, model_config: dict):
         super().__init__()
         self.hidden_dim = model_config["hidden_dim"]
