+++ models/tasks/__init__.py
@@ -1,0 +1,2 @@
+from .varmisuse import VarMisuseModel
+from .varnaming import VarNamingModel
