+++ scripts/test_varnaming.py
@@ -10,7 +10,7 @@ from data_processing.graph_var_miner.graph_var_miner_dataloader import (
 )
 from data_processing.vocabulary.great_vocabulary import GreatVocabulary
 from data_processing.vocabulary.spm_vocabulary import SPMVocabulary
-from models.varnaming import VarNamingModel
+from models.tasks.varnaming import VarNamingModel
 
 
 def main():
